Moved tag library to visionconstants
