+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/vision/ATVision.java
@@ -1,5 +1,7 @@
 package org.firstinspires.ftc.teamcode.vision;
 
+import static org.firstinspires.ftc.teamcode.vision.VisionConstants.getCenterStageTagLibrary;
+
 import android.util.Size;
 
 import com.acmerobotics.roadrunner.Vector2d;
@@ -114,41 +116,5 @@ public class ATVision {
                 tagPose.get(0) - y2,
                 tagPose.get(1) + x2
         );
-    }
-
-    // Position library credit to Michael from team 14343 (@overkil on Discord)
-    public static AprilTagLibrary getCenterStageTagLibrary() {
-        return new AprilTagLibrary.Builder()
-                .addTag(1, "BlueAllianceLeft",
-                        2, new VectorF(61.75f, 41.41f, 4f), DistanceUnit.INCH,
-                        new Quaternion(0.3536f, -0.6124f, 0.6124f, -0.3536f, 0))
-                .addTag(2, "BlueAllianceCenter",
-                        2, new VectorF(61.75f, 35.41f, 4f), DistanceUnit.INCH,
-                        new Quaternion(0.3536f, -0.6124f, 0.6124f, -0.3536f, 0))
-                .addTag(3, "BlueAllianceRight",
-                        2, new VectorF(61.75f, 29.41f, 4f), DistanceUnit.INCH,
-                        new Quaternion(0.3536f, -0.6124f, 0.6124f, -0.3536f, 0))
-                .addTag(4, "RedAllianceLeft",
-                        2, new VectorF(61.75f, -29.41f, 4f), DistanceUnit.INCH,
-                        new Quaternion(0.3536f, -0.6124f, 0.6124f, -0.3536f, 0))
-                .addTag(5, "RedAllianceCenter",
-                        2, new VectorF(61.75f, -35.41f, 4f), DistanceUnit.INCH,
-                        new Quaternion(0.3536f, -0.6124f, 0.6124f, -0.3536f, 0))
-                .addTag(6, "RedAllianceRight",
-                        2, new VectorF(61.75f, -41.41f, 4f), DistanceUnit.INCH,
-                        new Quaternion(0.3536f, -0.6124f, 0.6124f, -0.3536f, 0))
-                .addTag(7, "RedAudienceWallLarge",
-                        5, new VectorF(-70.25f, -40.625f, 5.5f), DistanceUnit.INCH,
-                        new Quaternion(0.5f, -0.5f, -0.5f, 0.5f, 0))
-                .addTag(8, "RedAudienceWallSmall",
-                        2, new VectorF(-70.25f, -35.125f, 4f), DistanceUnit.INCH,
-                        new Quaternion(0.5f, -0.5f, -0.5f, 0.5f, 0))
-                .addTag(9, "BlueAudienceWallSmall",
-                        2, new VectorF(-70.25f, 35.125f, 4f), DistanceUnit.INCH,
-                        new Quaternion(0.5f, -0.5f, -0.5f, 0.5f, 0))
-                .addTag(10, "BlueAudienceWallLarge",
-                        5, new VectorF(-70.25f, 40.625f, 5.5f), DistanceUnit.INCH,
-                        new Quaternion(0.5f, -0.5f, -0.5f, 0.5f, 0))
-                .build();
     }
 }
